refactor: consolidate duplicate SkeletonLoader into shared component
- Remove src/features/settings/components/shared/SkeletonLoader.tsx
- Update BillingTab.tsx and PayoutTab.tsx to import from
  src/shared/components/SkeletonLoader.tsx
- The shared version is a superset (adds variant, width, height props
  with same defaults), so all existing className-only usages are
  visually identical to before
- Update settings README to remove local SkeletonLoader entry and
  reference the shared implementation

Closes #473
